Move AboutPage component to pages directory
